requirements.txt
fastapi
uvicorn
sqlalchemy
databases
asyncpg
psycopg2
pandas
google-cloud-storage
scikit-learn